Driver/Volkslogger: convert_gcs() returns size_t

+++ src/Device/Driver/Volkslogger/vlapi2.cpp
@@ -283,7 +283,7 @@ VLAPI::read_igcfile(const TCHAR *filename, int index, int secmode)
 
   word serno;
   long sp;
-  long r = convert_gcs(0, outfile, logbuffer, 1, &serno, &sp);
+  size_t r = convert_gcs(0, outfile, logbuffer, 1, &serno, &sp);
   if (r > 0) {
     err = VLA_ERR_NOERR;
     print_g_record(outfile,   // output to stdout
